Refactor: Application, Game class 분리
