+++ src/main/java/baseball/Application.java
@@ -1,11 +1,10 @@
 package baseball;
 
+import game.Game;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Game game = new Game();
-        System.out.println("start!!");
-        game.set();
+        Game.set();
     }
 }
